Repocopy three more final ports for the complete MATE base:
 x11/gnome-panel -> x11/mate-panel
 x11/gnome-session -> x11/mate-session-manager
 x11-themes/gnome-themes -> x11-themes/mate-themes

I got it running by via ~/.xinitrc with a few of errors. The errors look like
going to be a bit hard to track down because the errors don't give very good
information. Joy. However, after seeing the Mate base running and it was
very clearly it is not good idea to have x11/mate to only have base MATE ports.
The x11/mate will have a full desktop with the MATE extras.


git-svn-id: svn://creme-brulee.marcuscom.com/ports-experimental/trunk@17448 df743ca5-7f9a-e211-a948-0013205c9059