Add webkitgtk 2.0.3.
Builds on current, GtkLaunch-1/3 (webkit test browser) and others launch but
crash. I'm not able to build this sucker yet with -g because my box with
4G RAM and 4G Swap dies with out of swap messages :)
9.x and 8.x need a new port and a patch to clang, I got both in development.
But even with those, build stops in gobject-introspection generation, yay.

But give other people a change to look at it on HEAD atleast for the crash bug.


git-svn-id: svn://creme-brulee.marcuscom.com/ports-experimental/trunk@18589 df743ca5-7f9a-e211-a948-0013205c9059